Implemented health percentage action

+++ Assets/Scripts/Domains/Health/Actions/HealthPercentageSO.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 namespace Domains.Health.Actions
 {
     [CreateAssetMenu(menuName = "BundleShop/Domains/Health/HealthPercentage", fileName = "HealthPercentage")]
-    public class HealthPercentageSO : BundleActionSO<float>
+    public class HealthPercentageSO : BundleActionSO<int>
     {
         public override Type DomainControllerType => typeof(HealthController);
     }
